+++ SuspiciousSeller/Assets/Scenes/PlayScene1.unity
@@ -135,6 +135,7 @@ GameObject:
   - component: {fileID: 311799850}
   - component: {fileID: 311799849}
   - component: {fileID: 311799852}
+  - component: {fileID: 311799853}
   m_Layer: 0
   m_Name: Enemy
   m_TagString: Untagged
@@ -206,8 +207,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c3605b38ddff5b74cb3e02edc472d695, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _movementSpeed: 2
-  _rotationSpeed: 0.2
 --- !u!4 &311799851
 Transform:
   m_ObjectHideFlags: 0
@@ -249,6 +248,19 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 4.801056, y: 3.7491102}
   m_EdgeRadius: 0
+--- !u!114 &311799853
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 311799848}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 56a6490ef5f1f34489644dc02b703efd, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _movementSpeed: 2
 --- !u!1 &519420028
 GameObject:
   m_ObjectHideFlags: 0
