+++ SuspiciousSeller/Assets/Scripts/NPC/NPCsDirectedMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 88c6128e4cdcca34db7f44e9b76eef5d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
