+++ SuspiciousSeller/Assets/Scripts/NPC/NPCsRandomMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 56a6490ef5f1f34489644dc02b703efd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
